+++ tests/testthat/test_pretty_output_functions.R
@@ -1,5 +1,7 @@
 context("pretty_output_functions")
 
+remove_ws <- function(x) gsub("\\s+", "", trimws(x))
+
 
 # test paste_tbl_grp
 test_that("paste_tbl_grp testing various options (no errors)", {
@@ -236,15 +238,24 @@ test_that("pretty_pvalues testing various options (no errors)", {
 
   # testing different outputs
 
-  expect_equal(object = pretty_pvalues(c(0.00000001, NA, 0.05, 1), digits = 3,
-                                       trailing_zeros = T, italic = T, output_type = "html"),
-               expected = c("<span style=\"  font-style: italic;   \" ><0.001</span>",
-                            "---", "0.050", "1.000"))
-  expect_equal(object = pretty_pvalues(c(0.00000001, NA, 0.05, 1), digits = 3,
-                                       trailing_zeros = T, bold = T,
-                                       italic = T, output_type = "html"),
-               expected = c("<span style=\" font-weight: bold; font-style: italic;   \" ><0.001</span>",
-                            "---", "0.050", "1.000"))
+  test_that("pretty_pvalues html output (italic)", {
+    expect_equal(
+      object = remove_ws(pretty_pvalues(c(0.00000001, NA, 0.05, 1), digits = 3,
+                                        trailing_zeros = T, italic = T, output_type = "html")),
+      expected = c("<spanstyle=\"font-style:italic;\"><0.001</span>",
+                   "---", "0.050", "1.000")
+    )
+  })
+
+  test_that("pretty_pvalues html output (bold + italic)", {
+    expect_equal(
+      object = remove_ws(pretty_pvalues(c(0.00000001, NA, 0.05, 1), digits = 3,
+                                        trailing_zeros = T, bold = T,
+                                        italic = T, output_type = "html")),
+      expected = c("<spanstyle=\"font-weight:bold;font-style:italic;\"><0.001</span>",
+                   "---", "0.050", "1.000")
+    )
+  })
 
   expect_equal(object = pretty_pvalues(c(0.00000001, NA, 0.05, 1), digits = 3,
                                        trailing_zeros = T, italic = T,