+++ NEWS.md
@@ -6,6 +6,7 @@
 * Update title and description in package DESCRIPTION file (#139)
 * Add `Median_Quartiles` to `pairwise_test_cont()` output (#99)
 * Improve documentation for datasets and fix typos in various documentation files (#148)
+* Add geometric mean, median, quantile, and standard deviation functions (#82)
 * Minor Changes:
   * Update `Overview()` to include new `Median_Quartiles` (#99)